+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - label displaying in BaseRadio 
 - support 'max' and 'min' attributes for BaseInput
 - ability to change placement in BaseRadio
+- new component BaseModalWindow
 
 ### Changed